requirements.txt
cryptography
Django==3.1.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
requests 
loguru
python-dotenv
opencv-python
math
time
django
djangorestframework
django-cors-headers
django-filter
loguru